Heurística inicial

+++ ListaAdjacenciaVet/header/grafo/Adjacencia.h
@@ -7,15 +7,15 @@ class Adjacencia
         int idVertice1; // Nó i pertencente a adjacência
         int idVertice2; // Nó j pertencente a adjacência
 
-        int weight; // Peso da aresta
+        float weight; // Peso da aresta
     public:
-        Adjacencia(int idVertice1, int idVertice2, int weight);
+        Adjacencia(int idVertice1, int idVertice2, float weight);
 
         int getIdVertice1();
         int getIdVertice2();
 
-        int getWeight();
-        void setWeight(int weight);
+        float getWeight();
+        void setWeight(float weight);
 };
 
 #endif // ADJACENCIA_H_INCLUDED
